pyproject.toml
[tool.poetry]
name = "samplefunctioncall"
version = "0.1.0"
description = ""
authors = ["abe <abe@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
openai = "^1.43.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
